Added save/load buttons and stuff

+++ NanoPores/distancetoatom.cpp
@@ -7,7 +7,7 @@
 #include <QDebug>
 #include <QElapsedTimer>
 
-float floatRandom(const int & min, const int & max) {
+float floatRandom(const float & min, const float & max) {
     static std::mt19937 generator;
     std::uniform_real_distribution<float> distribution(min, max);
     return distribution(generator);
